requirements.txt
certifi==2019.3.9
Django==2.2.2
get==2019.4.13
mysqlclient==1.4.2.post1
post==2019.4.13
public==2019.4.13
pytz==2019.1
query-string==2019.4.13
request==2019.4.13
sqlparse==0.3.0
